backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
moviepy==1.0.3
Pillow==9.5.0
edge-tts
aiofiles
python-multipart
